Les floats remplacent les doubles ! [18 mai 2022 à 20:22]

+++ libft/srcs/colors.c
@@ -6,13 +6,13 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/08 04:02:55 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/09 11:19:21 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/18 18:24:47 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	trgb_color(double t, double r, double g, double b)
+int	trgb_color(int t, float r, float g, float b)
 {
 	r = fmax(0, fmin(r, 255));
 	g = fmax(0, fmin(g, 255));
