+++ srcs/intersection/utils.c
@@ -6,27 +6,27 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/17 14:51:30 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/18 12:58:56 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/18 18:20:17 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
 /* Solution 1 : (-b - racine(delta)) / 2a */
-extern inline double	t1(double delta, double *abc)
+extern inline float	t1(float delta, float *abc)
 {
 	return ((-abc[1] - sqrt(delta)) / (2 * abc[0]));
 }
 
 /* Solution 2 : (-b + racine(delta)) / 2a
  * Racine de delta étant positif, on a t2 >= t1 */
-extern inline double	t2(double delta, double *abc)
+extern inline float	t2(float delta, float *abc)
 {
 	return ((-abc[1] + sqrt(delta)) / (2 * abc[0]));
 }
 
 /* En supposant que t1 < t2 */
-extern inline double	min_positive(const double t1, const double t2)
+extern inline float	min_positive(const float t1, const float t2)
 {
 	if (t1 > 0)
 		return (t1);
@@ -34,7 +34,7 @@ extern inline double	min_positive(const double t1, const double t2)
 }
 
 /* Position = origine + t * direction */
-extern inline t_v3	get_position(t_v3 origine, t_v3 direction, double t)
+extern inline t_v3	get_position(t_v3 origine, t_v3 direction, float t)
 {
 	return (add_vector(origine, mul_vector(direction, t)));
 }
